pyproject.toml
[project]
name = "uq-algorithm-selection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "keras>=3.11.3",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "tensorflow>=2.20.0",
    "tensorflow-probability>=0.25.0",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
]
